Now it's possible to send a message to a STOMP server with TSTompClient class.
Will use TStompClient class, that
will connect and publish a message to a STOMP server. I'm using the
TStompFrameBuilder to prepare a STOMP frame to send over the socket.

I'm encapsulating all the socket comunication from TStompClient with
TStompSocket, so I can make the verifications all in the TStompSocket
(TODO). Only the TStompSockerHarbour is implemented, still need to
implemente the TStompSocketADVPL.

The tests of TStompClient with use a TStompSocketFake version, to test
all returns of the TStompSocket.